requirements.txt
# RSS Feed Parsing
feedparser==6.0.10

# Claude API
anthropic==0.39.0

# Date Handling
python-dateutil==2.8.2

# Configuration Files
pyyaml==6.0.1

# HTTP Requests (for NYT API and potential web scraping)
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0
